Add joint restoration utility for pooled enemies
